SJRK-40: extract nodeModuleMounter to separate grade.
